backend/src/requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httpx[http2]
django-redis==5.4.0
requests
tzdata
django-cors-headers
Pillow>=10.0.0